folder structure adjustment and separated components containers page refactor + redux very first index.js setting complete
