+++ client/src/components/Layout.js
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import Header from "./Header";
 import Footer from "./Footer";
 
-const Layout = ({ children, setCardData, setCards, overflow, height }) => {
+const Layout = ({ children, setProducts, setCards, overflow, height }) => {
   const myList = [];
 
   return (
     <div style={{ overflow: overflow, height: height }}>
-      <Header setCardData={setCardData} setCards={setCards} />
+      <Header setProducts={setProducts} setCards={setCards} />
       {children}
       <Footer />
     </div>
