--- client/src/lib/createRequestSaga.js
@@ -1,20 +1,0 @@
-import { call, put } from "redux-saga/effects";
-import { startLoading, finishLoading } from "../store/modules/loading";
-
-export default function createRequestSaga(type, request) {
-  const SUCCESS = `${type}_SUCCESS`;
-  const FAILURE = `${type}_FAILURE`;
-
-  return function* (action) {
-    yield put(startLoading(type));
-
-    try {
-      const response = yield call(request, action.payload);
-      yield put({ type: SUCCESS, payload: response.data, meta: response });
-    } catch (e) {
-      yield put({ type: FAILURE, payload: e, error: true });
-    }
-
-    yield put(finishLoading(type));
-  };
-}
