+++ client/src/modules/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux";
+import counter from "./counter";
+
+const rootReducer = combineReducers({
+  counter,
+});
+
+export default rootReducer;
