+++ client/src/pages/LoginPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import LoginContainer from "../containers/LoginContainer";
+import Layout from "../components/Layout";
+
+const LoginPage = () => {
+  return (
+    <Layout>
+      <LoginContainer />
+    </Layout>
+  );
+};
+
+export default LoginPage;
