+++ client/src/pages/SearchPage.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Layout from "../components/Layout";
+import SearchContainer from "../containers/SearchContainer";
+
+const SearchPage = () => {
+  return (
+    <Layout>
+      <SearchContainer />
+    </Layout>
+  );
+};
+
+export default SearchPage;
